`atlas-syntax` removed
