+++ Cargo.toml
@@ -1,6 +1,5 @@
 [workspace]
 members = [
     "atlas",
-    "atlas-syntax",
     "atlas-misc",
 ]
