--- atlas-syntax/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "atlas-syntax"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ariadne = { version = "0.3.0", features = ["auto-color"] }
-atlas-misc = { path = "../atlas-misc" }
-
