--- atlas-syntax/src/ast/statements.rs
@@ -1,29 +1,0 @@
-use atlas_misc::span::WithSpan;
-
-use super::expr::{Expression, Identifier};
-
-#[derive(Debug)]
-pub enum Statement {
-    AssignmentStmt(Assignment),
-    IfStmt(If),
-    ExpressionStmt(Expression),
-    ReturnStmt(Return),
-}
-
-#[derive(Debug)]
-pub struct Assignment {
-    pub var_name: WithSpan<Identifier>,
-    pub value: WithSpan<Expression>,
-}
-
-#[derive(Debug)]
-pub struct Return {
-    pub value: Option<Expression>,
-}
-
-#[derive(Debug)]
-pub struct If {
-    pub cond: Expression,
-    pub body: Box<WithSpan<Statement>>,
-    pub else_body: Option<Box<WithSpan<Statement>>>,
-}
