+++ atlas/Cargo.toml
@@ -8,6 +8,5 @@ edition = "2021"
 [dependencies]
 ariadne = { version = "0.3.0", features = ["auto-color"] }
 atlas-misc = { path = "../atlas-misc" }
-atlas-syntax = { path = "../atlas-syntax" }
 clap = { version = "4.3.23", features = ["derive"] }
 heck = "0.4.1"
