+++ atlas/src/interfaces/lexer/errors.rs
@@ -1,4 +1,4 @@
-use crate::span;
+use crate::utils::span;
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum LexerError {
