+++ atlas/src/interfaces/lexer/mod.rs
@@ -1,16 +1,16 @@
 pub mod token;
 pub mod errors;
 
-use crate::span::WithSpan;
+use crate::utils::span::WithSpan;
 
 pub trait Lexer {
     fn new(file_path: String) -> Result<Self, std::io::Error>
         where Self: Sized;
-    fn tokenize(&mut self) -> Vec<WithSpan<token::TokenType>>;
-    fn check(&mut self, tokens: &[WithSpan<token::TokenType>]) -> Result<(), errors::LexerError> {
+    fn tokenize(&mut self) -> Vec<WithSpan<token::Token>>;
+    fn check(&mut self, tokens: &[WithSpan<token::Token>]) -> Result<(), errors::LexerError> {
         for token in tokens {
             match token.value {
-                token::TokenType::Unknown(c) => {
+                token::Token::Unknown(c) => {
                     return Err(errors::LexerError::UnknownToken(c, token.span))
                 }
                 _ => ()
