+++ atlas/src/interfaces/lexer/token.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, PartialEq)]
-pub enum TokenType {
+pub enum Token {
     //Keywords
     KwIf,       //if
     KwElse,     //else
@@ -91,83 +91,83 @@ pub enum TokenType {
 }
 
 
-impl ToString for TokenType {
+impl ToString for Token {
     fn to_string(&self) -> String {
         match self {
-            TokenType::KwIf => "if".to_string(),
-            TokenType::KwElse => "else".to_string(),
-            TokenType::KwFn => "fn".to_string(),
-            TokenType::KwInt => "int".to_string(),
-            TokenType::KwFloat => "float".to_string(),
-            TokenType::KwBool => "bool".to_string(),
-            TokenType::KwChar => "char".to_string(),
-            TokenType::KwString => "string".to_string(),
-            TokenType::KwVoid => "void".to_string(),
-            TokenType::KwReturn => "return".to_string(),
-            TokenType::KwTrue => "true".to_string(),
-            TokenType::KwFalse => "false".to_string(),
-            TokenType::KwLet => "let".to_string(),
-            TokenType::KwPub => "pub".to_string(),
-            TokenType::KwStruct => "struct".to_string(),
-            TokenType::KwEnum => "enum".to_string(),
-            TokenType::KwType => "type".to_string(),
-            TokenType::KwConst => "const".to_string(),
-            TokenType::KwPrint => "print".to_string(),
-            TokenType::KwNone => "none".to_string(),
-            TokenType::KwSelf => "self".to_string(),
-            TokenType::Ident(s) => s.to_string(),
-            TokenType::Int(i) => i.to_string(),
-            TokenType::Float(f) => f.to_string(),
-            TokenType::Char(c) => c.to_string(),
-            TokenType::String_(s) => s.to_string(),
-            TokenType::LParen => "(".to_string(),
-            TokenType::RParen => ")".to_string(),
-            TokenType::LBrace => "{{".to_string(),
-            TokenType::RBrace => "}}".to_string(),
-            TokenType::LBracket => "[[".to_string(),
-            TokenType::RBracket => "]]".to_string(),
-            TokenType::OpAdd => "+".to_string(),
-            TokenType::OpSub => "-".to_string(),
-            TokenType::OpMul => "*".to_string(),
-            TokenType::OpDiv => "/".to_string(),
-            TokenType::OpMod => "%".to_string(),
-            TokenType::OpEq => "==".to_string(),
-            TokenType::OpNe => "!=".to_string(),
-            TokenType::OpLt => "<".to_string(),
-            TokenType::OpLe => "<=".to_string(),
-            TokenType::OpGt => ">".to_string(),
-            TokenType::OpGe => ">=".to_string(),
-            TokenType::OpAnd => "&&".to_string(),
-            TokenType::OpOr => "||".to_string(),
-            TokenType::OpNot => "!".to_string(),
-            TokenType::OpAssign => "=".to_string(),
-            TokenType::OpAssignAdd => "+=".to_string(),
-            TokenType::OpAssignSub => "-=".to_string(),
-            TokenType::OpAssignMul => "*=".to_string(),
-            TokenType::OpAssignDiv => "/=".to_string(),
-            TokenType::OpAssignMod => "%=".to_string(),
-            TokenType::Colon => ":".to_string(),
-            TokenType::DoubleColon => "::".to_string(),
-            TokenType::Dot => ".".to_string(),
-            TokenType::DoubleDot => "..".to_string(),
-            TokenType::LArrow => "<-".to_string(),
-            TokenType::RArrow => "->".to_string(),
-            TokenType::FatArrow => "=>".to_string(),
-            TokenType::Pipe => "|".to_string(),
-            TokenType::Question => "?".to_string(),
-            TokenType::Semicolon => ";".to_string(),
-            TokenType::Comma => ",".to_string(),
-            TokenType::Ampersand => "&".to_string(),
-            TokenType::Unknown(c) => c.to_string(),
-            TokenType::UnterminatedString => "Unterminated string".to_string(),
-            TokenType::NewLine => "\n".to_string(),
-            TokenType::Tabulation => "\t".to_string(),
-            TokenType::EOF => "EOF".to_string(),
+            Token::KwIf => "if".to_string(),
+            Token::KwElse => "else".to_string(),
+            Token::KwFn => "fn".to_string(),
+            Token::KwInt => "int".to_string(),
+            Token::KwFloat => "float".to_string(),
+            Token::KwBool => "bool".to_string(),
+            Token::KwChar => "char".to_string(),
+            Token::KwString => "string".to_string(),
+            Token::KwVoid => "void".to_string(),
+            Token::KwReturn => "return".to_string(),
+            Token::KwTrue => "true".to_string(),
+            Token::KwFalse => "false".to_string(),
+            Token::KwLet => "let".to_string(),
+            Token::KwPub => "pub".to_string(),
+            Token::KwStruct => "struct".to_string(),
+            Token::KwEnum => "enum".to_string(),
+            Token::KwType => "type".to_string(),
+            Token::KwConst => "const".to_string(),
+            Token::KwPrint => "print".to_string(),
+            Token::KwNone => "none".to_string(),
+            Token::KwSelf => "self".to_string(),
+            Token::Ident(s) => s.to_string(),
+            Token::Int(i) => i.to_string(),
+            Token::Float(f) => f.to_string(),
+            Token::Char(c) => c.to_string(),
+            Token::String_(s) => s.to_string(),
+            Token::LParen => "(".to_string(),
+            Token::RParen => ")".to_string(),
+            Token::LBrace => "{{".to_string(),
+            Token::RBrace => "}}".to_string(),
+            Token::LBracket => "[[".to_string(),
+            Token::RBracket => "]]".to_string(),
+            Token::OpAdd => "+".to_string(),
+            Token::OpSub => "-".to_string(),
+            Token::OpMul => "*".to_string(),
+            Token::OpDiv => "/".to_string(),
+            Token::OpMod => "%".to_string(),
+            Token::OpEq => "==".to_string(),
+            Token::OpNe => "!=".to_string(),
+            Token::OpLt => "<".to_string(),
+            Token::OpLe => "<=".to_string(),
+            Token::OpGt => ">".to_string(),
+            Token::OpGe => ">=".to_string(),
+            Token::OpAnd => "&&".to_string(),
+            Token::OpOr => "||".to_string(),
+            Token::OpNot => "!".to_string(),
+            Token::OpAssign => "=".to_string(),
+            Token::OpAssignAdd => "+=".to_string(),
+            Token::OpAssignSub => "-=".to_string(),
+            Token::OpAssignMul => "*=".to_string(),
+            Token::OpAssignDiv => "/=".to_string(),
+            Token::OpAssignMod => "%=".to_string(),
+            Token::Colon => ":".to_string(),
+            Token::DoubleColon => "::".to_string(),
+            Token::Dot => ".".to_string(),
+            Token::DoubleDot => "..".to_string(),
+            Token::LArrow => "<-".to_string(),
+            Token::RArrow => "->".to_string(),
+            Token::FatArrow => "=>".to_string(),
+            Token::Pipe => "|".to_string(),
+            Token::Question => "?".to_string(),
+            Token::Semicolon => ";".to_string(),
+            Token::Comma => ",".to_string(),
+            Token::Ampersand => "&".to_string(),
+            Token::Unknown(c) => c.to_string(),
+            Token::UnterminatedString => "Unterminated string".to_string(),
+            Token::NewLine => "\n".to_string(),
+            Token::Tabulation => "\t".to_string(),
+            Token::EOF => "EOF".to_string(),
         }
     }
 }
 
-impl TokenType {
+impl Token {
     pub fn get_size(&self) -> usize {
         self.to_string().len()
     }
