+++ atlas/src/utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod case;
+pub mod span;
