+++ atlas/src/utils/span.rs
